Fixed bug: the ARA client did create routing table entries to its own location
The problem arose when a client started a route discovery by
broadcasting a FANT. This FANT would be rebroadcasted by the any
neighbor node in thus be received by the original source node.
The source node must ignore those packets instead of creating a
routing table entry to its own location.

I did also add a test to make sure this bug does not reappear in the
future :)